Version entire schema and provide initial migration plan
