+++ Vault/Sources/VaultFeed/Storage/Infrastructure/SwiftData/PersistedLocalVaultStoreFactory.swift
@@ -13,12 +13,12 @@ public final class PersistedLocalVaultStoreFactory {
             let storeURL = try getDocumentDirectory().appending(path: "PersistedLocalVaultStore-Main")
             let configuration = ModelConfiguration(
                 "PersistedLocalVaultStore",
-                schema: .init(versionedSchema: PersistedSchemaV1.self),
+                schema: .init(versionedSchema: PersistedSchemaLatestVersion.self),
                 url: storeURL
             )
             let container = try ModelContainer(
-                for: PersistedVaultItem.self,
-                migrationPlan: nil,
+                for: PersistedVaultItem.self, PersistedVaultTag.self,
+                migrationPlan: PersistedSchemaMigrationPlan.self,
                 configurations: configuration
             )
             return PersistedLocalVaultStore(modelContainer: container)
