+++ Vault/Sources/VaultFeed/Storage/Infrastructure/SwiftData/Schema/PersistedSchemaVersion.swift
@@ -1,0 +1,14 @@
+import Foundation
+import SwiftData
+
+/// The latest version of the schema.
+typealias PersistedSchemaLatestVersion = PersistedSchemaV1
+
+/// The most recent schema version for `PersistedVaultItem`.
+typealias PersistedVaultItem = PersistedSchemaLatestVersion.PersistedVaultItem
+/// The most recent schema version for `PersistedOTPDetails`.
+typealias PersistedOTPDetails = PersistedSchemaLatestVersion.PersistedOTPDetails
+/// The most recent schema version for `PersistedNoteDetails`.
+typealias PersistedNoteDetails = PersistedSchemaLatestVersion.PersistedNoteDetails
+/// The most recent schema version for `PersistedVaultTag`.
+typealias PersistedVaultTag = PersistedSchemaLatestVersion.PersistedVaultTag
